requirements.txt
# requirements.txt
# SCOPE framework dependencies

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
tqdm>=4.62.0

# LLM API clients
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
groq>=0.4.0

# Sentence embeddings
sentence-transformers>=2.2.0
torch>=1.9.0

# Utilities
python-dotenv>=0.19.0